3D UI Coins Implementation
In the 3D level, a UI has been implemented for the coins. In the upper left corner of the screen you can now see coins with No color. When the player touches a coin, then one of the coins in the UI will have color.

Also, completed the task of Spawning the last coin in the 3D area when you have 4 coins.